# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "API documentation for my application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/posts/{id}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a post",
                "description": "Update a specific post by ID.",
                "operationId": "106e2e3655fde890a91a8cd0c5d02615",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Post": {
                "required": [
                    "id",
                    "title",
                    "content",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "description": "Post ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Post title",
                        "type": "string"
                    },
                    "content": {
                        "description": "Post content",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UpdatePostRequest": {
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "description": "Updated post title",
                        "type": "string"
                    },
                    "content": {
                        "description": "Updated post content",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "Tymon JWT-Auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "security": [
                {
                    "bearerAuth": []
                }
            ]
        }
    ]
}